backend/infra/requirements.txt
# Observer-Eye Backend Dependencies
# Django Core
Django>=5.0,<6.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9

# Authentication
djangorestframework-simplejwt>=5.3.0
python-jose>=3.3.0

# Observability
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-django>=0.43b0
opentelemetry-exporter-otlp>=1.22.0
structlog>=24.1.0

# Utilities
python-dotenv>=1.0.0
celery>=5.3.6
redis>=5.0.1

# Production Server
gunicorn>=21.2.0
uvicorn>=0.27.0

# Development
pytest>=8.0.0
pytest-django>=4.7.0
factory-boy>=3.3.0
